Se cambia el tipo de dato al id de rol

+++ clicLava/src/main/java/org/clicLava/repository/RolRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RolRepository extends JpaRepository<Rol, Integer>{
+public interface RolRepository extends JpaRepository<Rol, Long>{
 	Optional<Rol> findByRol (String rol);
 }
